Make project script executable, portable
Hard coding /bin/bash limits the portability of shell scripts.
No features of bash are required for this script, so use the portable
POSIX shell /bin/sh instead.

Making the script executable allows it to be run directly: `./project1`

Renamed to follow the convention of executable programs not using an
extension on *nix systems.