Fix requirements.txt for problem1
The requirements.txt file for problem 1 was more of a README, so it has
been renamed to README.md and updated to reflect changes to the project.
A requirements.txt conforming to the pip requirements file format has
been added in its place.

The Python script for problem 1 is only compatible with Python 2.7, not
Python 3 or later.